==> Pipeline files/1_text_clean.ipynb <==
import pandas as pd

bucket = ""
import_name = "data/note_set.csv"
export_name = "data/note_set_cleaned.csv"

notes = pd.read_csv(f"s3://{bucket}/{import_name}")

#take a df series of text
#return text cleaned of unicode code points and bad characters that we have found in the notes
def clean_text(text):
    text = text.str.replace("<82>|<85>", ",", regex=True)
    text = text.str.replace("<92>", "'", regex=False)
    text = text.str.replace("<93>|<94>", '"', regex=True)
    #these two seem to mostly (though not always) appear in contexts that should have spaces
    text = text.str.replace("u00b7u00b0u00b7u00b0", "  ")
    text = text.str.replace("u00b7u00b0", "  ")
    text = text.str.replace("\\n", "  ", regex=True)
    #remove raw unicode
    text = text.str.replace("u0.b.|u2...|u0.a.|u00..|u00...|<..>", "", regex=True)
    #remove other misc. symbols + the strange $& & $!
    text = text.str.replace("\$\&|\$!|¼|½|·|`|°|•|§|—", "", regex=True)
    text = text.str.replace("_____-_____-_____", "")
    text = text.str.replace("+--------------+---------------+---------+---------------------+", "", regex=False)
    text = text.str.replace("_____", "")
    text = text.str.replace(" --   -- ", "")
    return text
    
notes["cleaned_notes"] = clean_text(notes.notes)
notes = notes.drop('notes', axis=1)
notes.to_csv(f"s3://{bucket}/{export_name}")
